Originating revision: r447 | cjplummer | 2006-11-16 22:52:14 -0800 (Thu, 16 Nov 2006) | 12 lines
Fixed Darwin JVMTI Support
--------------------------
                                                                                
Added some missing darwin JVMTI files.
                                                                                
Added a JVMTI bug fix to the darwin socket_md.c that was already in the linux version, and also enabled some code that was in a #ifdef __linux__.
                                                                                
When running jdwpgen, use -Xbootclasspath/p rather the -classpath, because the Mac OS version of JavaSE has a version of jdwpgen in the bootclasspath already, and we want to make sure that version doesn't get used. The linux JavaSE did not
have this problem.
                                                                                
For darwin builds, we need to add iconv to LINK_ARCH_LIBS. It is needed by jdwp. For linux, it seems to be part of glibc, so doesn't need to be explicitly linked with.
                                                                                
Reviewed by Bill.