Typo is fixed.
------------------------

!!! Note: comments for the previous putback are invalid, should be:

CR 6533824 (PCSL: long type overflow in pcsl_file_gettotalsize).

A complete solution for the issue is replace 'long' return value with 'jlong'.

jlong is good type for this purpose because it's machine independent and expanded to valid platform type. jlong is already used in pcsl interfaces like pcsl strings.

Merged -r3325:3336 with pcsl-cr-6533824.

Reviewed by Alexey Zavitaev and Sergey Troshin.