Added -XsplitVerify={true|false} option. Defaults to false for now, but will eventually default to true once the split verifier is better tested.
Fixed problem #1 in bug 6808728 in the split verifier. Use cb as the key instead of ClassID as the key in the StackMap cache,

Fixed problem #2 in bug 6808728 in the split verifier. Make sure StackMaps are removed from the cache once they are used.

Fixed problem #3 in bug 6808728 in the split verifier. It was not properly dealing with Object.clone() being called on an array type, which is a special case and does not occur in CLDC.

Reviewed by Dean